requirements.txt
psycopg2
pydantic
strawberry-graphql[cli]
fastapi
uvicorn

